development/projects/GPT-Agents/chapter_04/requirements.txt
crewai[tools]>=0.30.0
python-dotenv
langchain-openai
agentops
pyautogen
autogenstudio